tests/wpt/web-platform-tests/css/css-contain/container-queries/container-units-basic.html

<!doctype html>
<title>Container Relative Units: qi, qb, etc</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-lengths">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  .inline { container-type: inline-size; }
  .size { container-type: size; }
  .inline.outer { width: 500px; }
  .size.outer { height: 400px; }
  .inline.inner { width: 300px; }
</style>
<div id=ref></div>
<div class="inline outer">
  <div class="size outer">
    <div class="inline inner">
      <div id=child>Test</div>
    </div>
  </div>
</div>
<script>
  setup(() => assert_implements_container_queries());

  function assert_unit_equals(element, actual, expected) {
    try {
      element.style.padding = actual;
      ref.style.padding = expected;
      assert_equals(getComputedStyle(element).paddingLeft,
                    getComputedStyle(ref).paddingLeft);
    } finally {
      element.style = '';
      ref.style = '';
    }
  }

  test(function() {
    assert_unit_equals(child, '0qi', '0px');
    assert_unit_equals(child, '1qi', '3px');
    assert_unit_equals(child, '10qi', '30px');
    assert_unit_equals(child, '10qw', '30px');
    assert_unit_equals(child, '10qb', '40px');
    assert_unit_equals(child, '10qh', '40px');
    assert_unit_equals(child, '10qmin', '30px');
    assert_unit_equals(child, '10qmax', '40px');
  }, 'Container relative units');

  test(function() {
    assert_unit_equals(child, '10qi', '30px');
    assert_unit_equals(child, '10qb', '40px');
    assert_unit_equals(child, 'calc(10qi + 10qb)', '70px');
    assert_unit_equals(child, 'max(10qi, 10qb)', '40px');
  }, 'Container relative units in math functions');
</script>
